+++ spec/uglifier_spec.rb
@@ -22,7 +22,7 @@ describe "Uglifier" do
     it "raises an error when minifying JS with Harmony without harmony option" do
       source = "const foo = () => bar();"
       expect { Uglifier.new(:compress => false).compile(source) }
-        .to raise_error(Uglifier::Error)
+        .to raise_error(Uglifier::Error, /harmony/)
     end
   end